entity errors corected

+++ src/main/java/com/insurance/claim_service/entity/Claim.java
@@ -24,7 +24,7 @@ public class Claim {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "claim_id")
-    private Long claimId;
+    private Integer claimId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "policy_id", nullable = false)
